requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
dropbox==12.0.2
reportlab==4.2.2
python-dotenv==1.0.1
pydantic==2.9.2
